chore: componentized all the code to separate files
